# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  projects.ListProjectsResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      workplace_id:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
...
  tasks.GetTaskResponse:
    properties:
      assignee_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        type: string
      project_id:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  tasks.ListTasksResponse:
    properties:
      assignee_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        type: string
      project_id:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  tasks.UpdateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  tasks.UpdateTaskResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for DPC Tools.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DPC Tools API
  version: "1.0"
paths:
...
  /projects/{id}/tasks:
    get:
      description: Retrieve all tasks within a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/tasks.ListTasksResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Tasks
      tags:
      - tasks
...
  /tasks/{id}:
    get:
      description: Retrieve details of a specific task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/tasks.GetTaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update details of an existing task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated task details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/tasks.UpdateTaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update Task
      tags:
      - tasks
  /workplaces/{id}/projects:
    get:
      description: Retrieve all projects within a specific workplace
      parameters:
      - description: Workplace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/projects.ListProjectsResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Projects
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"